task 2.2 - refactor exceptions
